+++ src/productTypes/types/UnassignProductAttribute.ts
@@ -3,7 +3,7 @@
 // @generated
 // This file was automatically generated and should not be edited.
 
-import { AttributeTypeEnum, MeasurementUnitsEnum, WeightUnitsEnum } from "./../../types/globalTypes";
+import { AttributeTypeEnum, MeasurementUnitsEnum, AttributeInputTypeEnum, WeightUnitsEnum } from "./../../types/globalTypes";
 
 // ====================================================
 // GraphQL mutation operation: UnassignProductAttribute
@@ -43,6 +43,7 @@ export interface UnassignProductAttribute_productAttributeUnassign_productType_p
   filterableInDashboard: boolean;
   filterableInStorefront: boolean;
   unit: MeasurementUnitsEnum | null;
+  inputType: AttributeInputTypeEnum | null;
 }
 
 export interface UnassignProductAttribute_productAttributeUnassign_productType_variantAttributes {
@@ -55,6 +56,7 @@ export interface UnassignProductAttribute_productAttributeUnassign_productType_v
   filterableInDashboard: boolean;
   filterableInStorefront: boolean;
   unit: MeasurementUnitsEnum | null;
+  inputType: AttributeInputTypeEnum | null;
 }
 
 export interface UnassignProductAttribute_productAttributeUnassign_productType_weight {